+++ configs/datasets/DocILE-ELSA.yml
@@ -1,5 +1,5 @@
 dataset_name: DocILE-ELSA
 imdb_dir: /SSD/Datasets/DocILE/elsa_imdb
 # images_dir: /SSD/Datasets/DocILE/images
-images_dir: /SSD/Datasets/DocILE/images_dpi300
+images_dir: /SSD/Datasets/DocILE/images
 page_retrieval: Custom  # Oracle / Concat / Logits / Custom (for Hierarchical methods (Hi-LayoutLMv3, Hi-LT5, Hi-VT5) TODO: First
