+++ models/VT5.py
@@ -12,17 +12,17 @@ import transformers.models.t5.modeling_t5
 
 class ProxyVT5:
     def __init__(self, config):
-        self.batch_size = config['batch_size']
-        self.tokenizer = T5Tokenizer.from_pretrained(config['model_weights'])
-        self.model = T5ForConditionalGeneration.from_pretrained(config['model_weights'])
-        self.page_retrieval = config['page_retrieval'].lower() if 'page_retrieval' in config else None
-        self.max_source_length = config.get('max_source_length', 512)
+        self.batch_size = config.batch_size
+        self.tokenizer = T5Tokenizer.from_pretrained(config.model_weights)
+        self.model = T5ForConditionalGeneration.from_pretrained(config.model_weights)
+        self.page_retrieval = config.page_retrieval.lower() if 'page_retrieval' in config else None
+        self.max_source_length = getattr(config, 'max_source_length', 512)
 
-        t5_config = CustomT5Config.from_pretrained(config['model_weights'])
-        t5_config.visual_module_config = config['visual_module']
+        t5_config = CustomT5Config.from_pretrained(config.model_weights)
+        t5_config.visual_module_config = config.visual_module
 
-        self.spatial_embedding = SpatialEmbeddings(t5_config).to(config['device'])
-        self.visual_embedding = VisualEmbeddings(t5_config).to(config['device'])
+        self.spatial_embedding = SpatialEmbeddings(t5_config).to(config.device)
+        self.visual_embedding = VisualEmbeddings(t5_config).to(config.device)
 
     def parallelize(self):
         self.model = nn.DataParallel(self.model)
